+++ src/main/java/pl/plajer/villagedefense/Main.java
@@ -69,7 +69,6 @@ import pl.plajer.villagedefense.handlers.language.Messages;
 import pl.plajer.villagedefense.handlers.powerup.PowerupRegistry;
 import pl.plajer.villagedefense.handlers.reward.RewardsFactory;
 import pl.plajer.villagedefense.handlers.setup.SetupInventory;
-import pl.plajer.villagedefense.handlers.setup.SetupInventoryEvents;
 import pl.plajer.villagedefense.kits.KitManager;
 import pl.plajer.villagedefense.kits.KitRegistry;
 import pl.plajer.villagedefense.kits.basekits.Kit;
@@ -229,7 +228,6 @@ public class Main extends JavaPlugin {
     KitRegistry.init(this);
     new SpectatorEvents(this);
     new QuitEvent(this);
-    new SetupInventoryEvents(this);
     new JoinEvent(this);
     new ChatEvents(this);
     setupPluginMetrics();